requirements.txt
Django==1.9.6
Pillow==3.2.0
pytz==2016.4
django-taggit==0.19.1
sorl-thumbnail==12.4a1
requests==2.10.0
amqp==2.0.1
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
kombu==3.0.35
django-celery==3.1.17
vine==1.1.0
django-redis-sessions==0.5.6
tornado==4.4.1
python-social-auth==0.2.21
simplejson==3.8.2
django-extensions==1.7.3
requests-facebook-py3==0.2.1
vk==2.0.2
django-filebrowser-no-grappelli==3.6.2
django-tinymce==2.4.0
Django-tinymce-filebrowser==0.2.1
django-parler==1.6.5
django-social-share==0.3.1